Added new test to verify negative results of clEnqueueSignalSemaphoresKHR
