+++ test_conformance/extensions/cl_khr_semaphore/CMakeLists.txt
@@ -3,7 +3,7 @@ set(MODULE_NAME CL_KHR_SEMAPHORE)
 set(${MODULE_NAME}_SOURCES
          main.cpp
          test_semaphores.cpp
-         test_semaphores_negative_wait.cpp
+         test_semaphores_negative_wait_signal.cpp
          semaphore_base.h
 )
