+++ test_conformance/extensions/cl_khr_semaphore/main.cpp
@@ -44,6 +44,16 @@ test_definition test_list[] = {
                      Version(1, 2)),
     ADD_TEST_VERSION(semaphores_negative_wait_invalid_event_status,
                      Version(1, 2)),
+    ADD_TEST_VERSION(semaphores_negative_signal_invalid_command_queue,
+                     Version(1, 2)),
+    ADD_TEST_VERSION(semaphores_negative_signal_invalid_value, Version(1, 2)),
+    ADD_TEST_VERSION(semaphores_negative_signal_invalid_semaphore,
+                     Version(1, 2)),
+    ADD_TEST_VERSION(semaphores_negative_signal_invalid_context, Version(1, 2)),
+    ADD_TEST_VERSION(semaphores_negative_signal_invalid_event_wait_list,
+                     Version(1, 2)),
+    ADD_TEST_VERSION(semaphores_negative_signal_invalid_event_status,
+                     Version(1, 2)),
 };
 
 const int test_num = ARRAY_SIZE(test_list);
